Refactor Nav Component
