+++ src/app/constant/nav.ts
@@ -1,0 +1,2 @@
+export const borderCss =
+  "1px 0 grey, -1px 0 grey, 1px 0 grey, -1px 0 grey, 1px 1px black, -1px -1px black, 1px -1px black, -1px 1px black";
